added mastercllock and changes to makefile so that it compiles in Windows

+++ entities.cpp
@@ -6,19 +6,19 @@ Entity::Entity(){}
 
 Entity::~Entity(){}
 
-Entity::Entity(float& x_in, float& y_in): x(x_in), y(y_in){}
+Entity::Entity(int& x_in, int& y_in): x(x_in), y(y_in){}
 
 
 // Projectile class
 Projectile::Projectile():Entity(){}
 Projectile::~Projectile(){}
-Projectile::Projectile(float& x_in, float& y_in):Entity(x_in,y_in){}
+Projectile::Projectile(int& x_in, int& y_in):Entity(x_in,y_in){}
 
 
 // Rocket class
 Rocket::Rocket():Projectile(){}
 Rocket::~Rocket(){}
-Rocket::Rocket(float& x_in, float& y_in):Projectile(x_in,y_in){}
+Rocket::Rocket(int& x_in, int& y_in):Projectile(x_in,y_in){}
 
 void Rocket::move(){y-=step;}
 
@@ -26,7 +26,7 @@ void Rocket::move(){y-=step;}
 // Bomb class
 Bomb::Bomb():Projectile(){}
 Bomb::~Bomb(){}
-Bomb::Bomb(float& x_in, float& y_in): Projectile(x_in,y_in){
+Bomb::Bomb(int& x_in, int& y_in): Projectile(x_in,y_in){
 	y += Engine::SpriteSize;
 }
 
@@ -36,14 +36,14 @@ void Bomb::move(){y+=step;}
 // Ship class
 Ship::Ship():Entity(){}
 Ship::~Ship(){}
-Ship::Ship(float& x_in, float& y_in):Entity(x_in, y_in){}
+Ship::Ship(int& x_in, int& y_in):Entity(x_in, y_in){}
 
 
 // Alien class
 
 Alien::Alien():Entity(){}
 Alien::~Alien(){}
-Alien::Alien(float& x_in, float& y_in):Entity(x_in, y_in){}
+Alien::Alien(int& x_in, int& y_in):Entity(x_in, y_in){}
 
 bool Alien::Shoot(std::vector<Bomb>& bombs)
 {
